Update move controls so that react doesn't throw a warning. It doesn't like that the child is changing state in the parent.
